refactor(hosted): name the Responses input admission module for what it admits
