+++ packages/hosted/src/index.ts
@@ -1,18 +1,4 @@
 export {
-  admitBrainInput,
-  admitBrainInputItem,
-  brainOutputReplayable,
-  maximumHostedBrainInputItems,
-  maximumHostedBrainRequestBytes,
-  RESPONSES_CALLER_TYPE,
-  RESPONSES_CONTENT_PART_TYPE,
-  RESPONSES_INPUT_ITEM_TYPE,
-  RESPONSES_ITEM_STATUS,
-  RESPONSES_MESSAGE_PHASE,
-  RESPONSES_MESSAGE_ROLE,
-  serializedRequestBytes,
-} from "./brain-admission.js";
-export {
   BRAIN_TURN_AUTHORITY,
   type BrainTurnAuthority,
   brainTurnAuthorityFromWire,
@@ -72,3 +58,17 @@ export {
   type RealtimeCredential,
   realtimeCredentialIsUsable,
 } from "./realtime-contract.js";
+export {
+  admitBrainInput,
+  admitBrainInputItem,
+  brainOutputReplayable,
+  maximumHostedBrainInputItems,
+  maximumHostedBrainRequestBytes,
+  RESPONSES_CALLER_TYPE,
+  RESPONSES_CONTENT_PART_TYPE,
+  RESPONSES_INPUT_ITEM_TYPE,
+  RESPONSES_ITEM_STATUS,
+  RESPONSES_MESSAGE_PHASE,
+  RESPONSES_MESSAGE_ROLE,
+  serializedRequestBytes,
+} from "./responses-input.js";
