+++ packages/hosted/src/responses-input.ts
